requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.15,<4.0
channels>=4.1,<5.0
channels-redis>=4.2,<5.0
redis>=5.0,<6.0
psycopg[binary]>=3.2,<4.0
gunicorn>=22.0,<23.0
stripe>=11.0,<12.0
Pillow>=10.0,<11.0
django-csp>=4.0,<5.0
celery>=5.4,<6.0
